backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "biopython>=1.86",
    "fastapi>=0.133.0",
    "google-genai>=1.33.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "reportlab>=4.4.10",
    "uvicorn[standard]>=0.41.0",
]
